camlistore/ui: add dialog to import shared items
This change adds an 'Import share' entry in the main menu bar. This
entry creates a dialog that allows the user to enter the URL of a share
claim, and to start importing the blobs shared through said claim. All
shared blobs found are forwarded as they are to the blobserver that the
web UI is configured for. Directory entries from directory blobs are
followed to discover the directory children. Same story for file
contents referenced by file blobs. Only directories and regular files
are supported for now, but adding the other types should be trivial as
most of the code already exists in camget.

One possible optimization would be to add a handler to the Camlistore
server that would do most of the work that this changes does client-side
(i.e. fetching blobs from the share blob). Then the client side would
only have to request the server to do so on the given share URL. The
obvious win is that the [shared blob]->[our server] network path is
probably faster than the [shared blob]->[our client]->[our server] one.
However, I'm not sure it fits the role of a server so I think it's ok to
start by having this feature live mostly on the client side for now.

Change-Id: I94cd03f834d7d1e9f4ec5c016ffdf16d911856c5