server/camlistored/ui: revert ref predicate to constraint
The Blob aspect in the web UI for non-permanodes is broken, because the
initial search now uses a ref: predicate instead of a Constraint. That
is because for now, expression searches (as opposed to constraint searches)
only match permanodes, so the initial search on the concerned blob
fails. See issue #944.

Until we figure out whether we want to allow non-permanode expression
searches, this change therefore reverts that initial search to a constraint
instead of a ref: predicate search.

And since the getQueryExprOrRef function in search_session.js still uses
a ref: predicate, the map aspect (for the single node case) is not
affected.

Issue #944

Change-Id: I5c4b4d0bca704a550b6a730f24b8e96568c5c8b6