server/camlistored/ui: add read-only mode to web UI
This change allows to basic auth login on the web UI with read-only
credentials, so that all of the contents can be viewed/searched etc, but
not modified in any way.

Will split in 4 different commits later, because that change involves:

1) adding the read-only mode to pkg/auth/
2) using the read-only mode in the web UI
3) adding a logout feature, so it's easy to switch between read-only and
normal mode.
4) split the gopherjs "goreact" namespace into distinct ones, which should be
completely orthogonal to the rest of this change.

Change-Id: Ie6219247fbe1b1c957125193d9d7515d0ee62d38